refactor: menugroup 도메인과 menu 도메인 통합
