+++ src/main/java/kitchenpos/menu/domain/MenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menugroup.domain;
+package kitchenpos.menu.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
