+++ src/main/java/kitchenpos/menu/dto/MenuGroupRequest.java
@@ -1,21 +1,25 @@
-package kitchenpos.menugroup.dto;
+package kitchenpos.menu.dto;
 
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
-public class MenuGroupResponse {
+public class MenuGroupRequest {
     private Long id;
     private String name;
 
-    public MenuGroupResponse() {
+    public MenuGroupRequest() {
     }
 
-    public MenuGroupResponse(Long id, String name) {
+    public MenuGroupRequest(Long id) {
         this.id = id;
+    }
+
+    public MenuGroupRequest(String name) {
         this.name = name;
     }
 
-    public static MenuGroupResponse of(MenuGroup menuGroup) {
-        return new MenuGroupResponse(menuGroup.getId(), menuGroup.getName());
+    public MenuGroupRequest(Long id, String name) {
+        this.id = id;
+        this.name = name;
     }
 
     public MenuGroup toMenuGroup() {
