+++ src/main/java/kitchenpos/menu/dto/MenuGroupResponse.java
@@ -1,25 +1,21 @@
-package kitchenpos.menugroup.dto;
+package kitchenpos.menu.dto;
 
-import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
-public class MenuGroupRequest {
+public class MenuGroupResponse {
     private Long id;
     private String name;
 
-    public MenuGroupRequest() {
+    public MenuGroupResponse() {
     }
 
-    public MenuGroupRequest(Long id) {
-        this.id = id;
-    }
-
-    public MenuGroupRequest(String name) {
-        this.name = name;
-    }
-
-    public MenuGroupRequest(Long id, String name) {
+    public MenuGroupResponse(Long id, String name) {
         this.id = id;
         this.name = name;
+    }
+
+    public static MenuGroupResponse of(MenuGroup menuGroup) {
+        return new MenuGroupResponse(menuGroup.getId(), menuGroup.getName());
     }
 
     public MenuGroup toMenuGroup() {
